Can evolve levels with Mario GAN #283
Took the CPPN Mario Level evolution and split it into two tasks, one of which is the new Mario GAN. Some code is shared, but now there are separate tasks for each type of evolution. GAN latent vector evolution uses the BoundedRealValuedGenotype, and there were some minor issues with making that work.