Mario GAN Model is now a parameter #283
Renamed the Mario GAN model to a name that makes it clear it is from GECCO 2018, and then made the model name a command line param so that different models can be tried.

Long term goal is to bring in the extra models that Vanessa trained with different level groups, but those also used a different tile encoding, so I need to enable and parameterize that too.