#283 Scaled progress fitness and uncovered bugs
There are two bugs here. One is that for MarioGAN the measure of level completion is off, so that agents that complete the levels do not receive a fitness indicating that they have beaten the level. However, even ignoring this, jumps seem to be incorrectly tallied in both Mario GAN and Mario CPPN. Not all jumps count for some reason.

Need to fix these bugs