#283 Fitness fixed?
The level percentages are now correct. CPPNs and the GAN now define the length of the level differently in each task, which makes it easier to calculate the fitness correctly.

I also changed how the jump count is computed, by copying over code from Vanessa's Games Benchmark branch of the MarioGAN code. Now it almost seems as though too many jumps are being counted, though I'm not entirely clear on the subtleties of the calculation ... however, I'm willing to trust the code as it is.

Next step is to bring in Vanessa's extra tile encodings and the other GANs she trained (for over and under worlds)