#283 MarioGAN works in Windows now!
This is because the latest version of Pytorch finally works in Windows
10. However, apparently PyTorch disallowed the use of "." in model names
a couple of versions ago, which means that the models that Vanessa and
Sebastian trained are deprecated.

In order to deal with this, some fixes had to be made to the code. First
the names in the models had all "." replaced with ":". Then, when the
models are loaded, the OrderedDict that represents the saved model has
to have its keys changed. More accurately, a new OrderedDict is made
with values in the same order, but with different (valid) key names.