#283 Removed most CPPN interface elements
MarioGAN evolution does not involve CPPNs, so now the reference to
actuvation functions is gone, along with the network view buton. The
other buttons seem to work, although there is still a slider for level
length that needs to be removed/replaced with one for chunk count.

Also need to fix display scaling still.