Batch file to launch a human player straight into the game
This bypasses the menu and has the human player immediately join a game on a specific IP address (currently the local host). This seems to work fine if I manually launch the server from a batch file, but doesn't work yet when MM-NEAT launches the server, which is strange. Needs some troubleshooting.