#291 Successful ZeldaGAN Training
Was able to train a GAN on Zelda rooms from the first dungeon. I'm committing the trained model and fake samples. The samples indicate that there is perhaps some overfitting ... the GAN recreates the levels too well, but we want variation.

Will probably eventually delete of move these files elsewhere, but I wanted to save them for now.

I also verified that generator_ws can be used with this new trained model, but the next step is to take its output into Java and play a level with GVG-AI