#291 Successfully load and play ZeldaGAN levels
The python files properly crop the GAN output based on type now (using
the number of tile types as a proxy to determine which game is being
played). Java can send a latent vector to the GAN to get a Zelda room
and then play it.

This also revealed to me that the GAN (even for Mario) was outputting to
a space larger than is needed for the level. Then, the excess is
cropped. A bit strange, but this creates some flexibility actually.

Next: ZeldaGAN is generating some new content, but train models with
smaller latent vector to discourage overfitting. Then share results!