Problem launching only native bots
I created a way to launch a server with native bots on different teams, but the severa immediately exits because it normally only stays running as long as a gamebots agent thread is running. Since none are launched, a new way needs to be devised to keep the server running for the duration of a match.