Progress in launching experiments and trials from command line
A batch file both starts the bot in MMNEAT and launches the UT client to play the game, but I still have the problem of not connecting to the server. Also need to figure out what happens when the time limit runs out ... bots probably just disappear. Need to change that.