Still tweaking
The utEvolvingBotSkill still needs to be tweaked. At a value of 5, Jude does worse in a 1 vs 2 match (hardly surprising) but I need to see how the team behaviors look. More importantly, I need to see how Ethan does, but this can only happen on the ToP.

I also need to tweak the server time out. The current set up is much better, but the bot still disappears right before the end of the match. I just made the buffer time higher, but don't know if it is enough yet. I feel like getParams().getEvalSeconds()  may not be returning the value I expect. I suppose the other possibility is that bot initialize happens significantly before the server actually launches the match, which actually might be possible.