Ethan skill and instructions
Ethan skill boosted to 7, since this is what Ethen evolved with. The point is for Ethan to be a very good player, but not a good teammate.

The other tweak is to the instructions. I added warnings about fixing the full screen problem, but it inexplicably fixed itself! Weird. Hopefully, it will stay fixed.