Comments added; code tangentially modified
The only change I made to any of the files, besides commenting on all of them and importing the other three scenarios into VizDoom, involved deleting the "setScaleFactor(1.0);" line in the codes of certain scenarios. Since the scale factor was already initialized to 1.0 in VizDoom, there was no purpose in keeping a line that would, essentially, "set" it to the same value.