All the Zentangle code present
However, if you run this it will not work because it tries to save the results in a directory that does not exist. The simple solution is to simply create that directory, but I want to make things a bit less error prone and have the code create the save directory, which will take a bit of fiddling to do right.