Clean up unnecessary array code
The original code made entries in these arrays for each of four rotations of a tile, but WFC handles that on its own. This code is putting items in these arrays that are never used, so I simply removed them. Of course, this leaves me with an array that is always size 1, which makes no sense ... for the moment