#321 Improved point spaceing, but removed sig diffs from bottom
Wanted to plot significant differences between conditions at the bottom of the plot. However, I was reminded that all such comparisons are actually two-way (A vs B not always same as B vs A), and that even publications from Clune's group generally don't show all comparisons. It makes more sense to select certain conditions to do pairwise comparisons on.

Therefore, the code is disabled (commented out) and can be easily tweaked in the future for specific data sets