Generalized A* Search Code and added other search methods
A* is basically the same as UCS and BFS, except for the use of a Priority Queue with a heuristic. I generalized the code to get the appropriate data structure from an abstract method, so that the same code could be used for UCS and BFS.

Could also add DFS with a stack, but it seems unlikely that I would need that.