Commnets for guidance on #361
I didn't fix the issue, but I have some guidance on how to do so. You need to add a pause that shows the dungeon with all of the searched states X-ed out on each iteration. Additionally, you need to figure out if the state returned by makePlayable is the one it is supposed to be ... should be a state that does NOT have an X but is right next to one ... one of the states where you removed a wall. @jakeryang