Fixed Maven build issue
For a long time, Maven has been complaining that it can't find Action even though it is clearly defined and implemented by GridAction in ZeldaState.

After looking around a lot I found that Maven for some reason doesn't like it when you import into a class contents of that same class in order to shorten names. Therefore, I changed this code.

It makes referring to the directions a bit more tedious, but should be equivalent to what you had before @jakeryang . It successfully Maven builds, as long as I used -DskipTests ... there is a test failing which I'll open an issue about soon.