Trying to set up some new Tetris experiments
So far, these are just copies of selected batch files from the last two papers, renamed with a consistent convention. Also made it so holes are a +1 input instead of -1 so that I can switch to ReLU, though that is not used in these batch files yet. Just need to make sure they seem to run properly first. Then, next step will be to add SZ Tetris, and also maybe add another architecture.