Batch files will record play sessions now
Annoyingly, they do not close at the end of the session. We had the same problem with UT2004 and never solved it then, but I'll see if we can find a way this time.

Also, we will need to change the recording window to match the resolution on whichever machine we end up conducting the study on. That will wait until next week.

#367