VB Script file can terminate ffmpeg
This seems to work. The VBScript program sends the key press q to the program which stops the recording elegantly without chopping anything off ... though that might also be due to the pause. I'm going to try removing that next, but wanted to commit this first becuase it does work. #367