Troubleshooting inconsistent levels
Different levels are produced by batch file with jar and from running in Eclipse. When a random seed of 4 is used, this is a problem, because there is an unsolvable block puzzle we're trying to fix ... however, the level only appears when launched from a jar file and not from Eclipse. Why not? #410