Random Strings
I think the random Strings being generated could be the source of the problem we are seeing. We already know the Strings are different each time. They are also used to store things in a HashMap, and I think that the String affects the order of iteration, which could cause different outcomes. We'll see