Added code for debugging random numbers
If RANDOM_DEBUG is set to true, then a lot of information will be printed every time the random generator is used. Currently set to false.

Also, the reason I added a commented line to MMNEAT is because uncommenting this will launch the code in a way that can be compared with Maven output. It is important to launch the code in Eclipse from MMNEAT so that the stack traces will line up when comparing with running from Maven.