Fixed questionable use of generic Exception
Don't throw generic Exception @jakeryang . Always be specific. In this case, I found the IllegalStateException to be most appropriate, which is also nice because it is unchecked, and doesn't need to tag every method in the call chain with "throws Exception". The other exception that was coming up was FileNotFound, but this is better handled by a try-catch in the method itself that exits immediately on failure.