Problem is movable blocks
#421 The problem with the novelty calculation seems to be that movable blocks in the Dungeon version of the representation simply show up as FLOOR Tiles. In the text files, they are represented as BLOCKs.

I was surprised that the movable blocks didn't show up as one of MOVABLE_BLOCK_UP/DOWN/LEFT/RIGHT, since I could address that issue in the novelty calculation.

However, as is, I need @jakeryang  to make sure that before the Dungeons are sent to the novelty calculation, any movable blocks are transformed to BLOCK in the representation.

Perhaps in DungeonNovelty.averageDungeonNovelty a method call can be made on the Dungeon before using getNodes