Solved problem with too many enemies
I believe this solves #453 and I hope it doesn't create any new errors. This error also requires some explanation.

Originally, the GAN was treating an int code of 2 as P (from VGLC) for water. However, this version of the GAN was generating int codes that were being ignored, so for the GECCO papers I trained a new GAN that actually had an output layer that generated only 3 possible int codes.

However, the in-game code of 2 is actually treated as an enemy creature. So, the reason there were too many enemies was that every occurrence of a water tile was being rendered as an enemy.

Basically, this error was only affecting the levels that did NOT use the GAN, since when I tweaked the GAN encoding I also fixed its mapping (this happens in the generateOneRoomListRepresentationFromGAN method of ZeldaGANUtil). However, by mapping P to 5 instead of 2, the other levels should behave as well.

Unfortunately, the random number generator must have been affected at some point, so that the interiors of rooms are no longer the same as in the original subject study (see https://people.southwestern.edu/~schrum2/SCOPE/zelda-graphgan.php )