Fixed up and down door issue
#443
Fixed the issue where only one of the door tiles were being removed from the up and down doors by changing the if statement in the loop to a while loop.

This ensures the whole door is removed in hte GAN representation on the console.

Makes it so there are no doors on the outside layer anymore.