Greatly simplifying code, but still not working
I'm committing this code, but I plan to completely change it again and abandon/completely transform this issue. Calling the search algorithm over and over is simply too computationally expensive. Instead of placing the keys "intelligently" we'll settle for placing them sparsely. The level might not be beatable, but we won't have an overabundance of keys. #446