Don't count locked doors in unreachable rooms #446
Although the total number of locked doors still includes unreachable rooms, during the actual key placement process, the code double checks if the room is reachable. If not, then the locked doors are subtracted from the total count.

This leads to reasonably palcement of keys across the level. It is sparse, but there are enough keys. Sometimes the keys will not be in the right place to allow successful completion of the level, especially in dungeons with few locked doors, but that's fine.