Resolved #465
Now, if one goes through a recently unlocked puzzle door into a room with an unmoved puzzle block, the door back will be locked until that puzzle block is also pushed. At the same time, levels with a one-way flow in which the room after the puzzle locked door does not have a puzzle block will remain unlocked.