#441 Distinct rooms are now a part of the fitness
However, I have not been able to generate a dungeon wherein there are indistict rooms, so I'm not totally sure that the counter is accurate. So far the numer of reachable rooms has always been equal to the number of distinct rooms. I have checked 10 of these dungeons and each reachable room is distinct.