Started code to render levels from the GAN model
#394
Added a parameter to get the model of the GAN that we trained on Lode Runner.
Added a case to allow for lode runer levels to be rendered from the GAN.
Updated generator_ws.py to default to lode runner when it is 8 tiles.
Added a class called LodeRunnerGANUtil to help control the rendering of levels from the GAN.