#457 Room adjacencies needed
Although we don't want to add connections between rooms simply because they are adjacent, we do need to connect rooms if there are doors. However, simply having the doors appear in the List of Lists representation is not enough. The rooms need to be expicitly connected during dungeon creation. This should be done conditionally based on whether there is a door available. @cappsb