ZeldaGANLevelBreeder now has the correct dimensions for every model
#480
The levels with 3 tile types now a have the correct width and height. Had to add a new condition parameter for selcting the height and width in the GANProcess.java file.
Also cleaned up some constants for width and height of LodeRunner levels so they are only in one place. #478
#478 Added comments into this class. Javadocs on all the methods and for the class itself