Trained new model on only the first 5 levels
#394
The TEMP models are from the version of training that were trained off of all 150 levels.
Renamed model names to end with "iterations_latentvectorsize_number of tiles"
Added JSON training data for the new models
Added a parameter to allow for multiple models to be used in  the level breeder. #483
Started to write code to make the levels playable using hte IceCreamYou repo that we have already made work on its own. #482 #472