Added new trainign sets and made it possible to pick different model with differnet tile encondings
#394
Added new training sets for the first 20 levels and the first 50 levels with both 6 and 7 tiles mappings.
#483
Added updated parameter with the correfct file name, I chagned it the file names to make it easier to distinguish between them.
#478
Also makes it easier to be able to switch between models in the level breeder.