Improved the scaling issue
#472
The scaling is much better, but still needs to be improved. I took the size of the ice cream you levels and divided it with the width and height of our levels. THis made thins show up in the play window and in hte right place for the most part.
There is still a problem with the placement of the enemies for some reason where they all show up on the bottom.
Also, the player gets placed and then can only move limited amounts because I think they might get stukc in the ground because of the placement.