#495 possible progress. Added new log
With the same general format as Archive, I created a new Integer[] called eliteProper that is filled with -1's and replaces those withnumbers based on what is and is not a CPPN or Direct. This will probably fail because I used i++ after each iteration of the population. This implies that the index will be off (I believe)