LodeRunnerState and LodeRunnerAction
#474
Created the LodeRunnerState class that allows the A* algorithm to play the level by defining the moves you can make at any given time, and priovdes a way to evaulate winning a level.
Inside this class, there is a sub class called LodeRunnerAction that defines all of the actions that a player can do as they play a level of LodeRunner.