Modified contents and file names of Mario batch files
#442 @cappsb I changed some settings in these files, and also changed the names. I will probably change them some more in a moment.

Important: You will need to make a similar batch file for CPPNThenGAN in Mario #488 . When you do that, be sure to base that batch file off of these. Any setting shared in the CPPN and Direct batch files for Mario should also be in the CPPNThenDirect batch file you make. Most additional settings from one or the other should also be included. Also compare with the Zelda CPPNThenGAN batch file to see how it differs from its counterparts.