Started on vizualizing the path A* takes
#474
Made it so it draws a line from 5,5 to 10,10 to start getting an idea of how to create the path marked with x's.
made a new createBufferedImage method in LodeRunnerRenderUtil that does not need an array of bufferedImages for the tiles.
Moved teh LodeRunnerLevelTask class to the LodeRunner from the loderunner.astar package to the loderunner package