Started working on digging actions
#474
Added a hashset to keep track of the holes that hacve been dug.
Added it to the constructors so we can use it with each state.
added optons for DIG_RIGHT and DIG_LEFT as actions for lode runner player.
Added code in each of those cases to start adding the points to the set if they are diggable ground.
Still not done, but a good start.