#474 Removing digging actions
They are too costly to apply A* to. However, if we simply pretend it is possible to move down through diggable blocks, then we can still find solution paths. This is a lot of progress, but it still fails if there are two layers of diggable blocks to go through.