LodeRunnerVGLCUtil launches level1 into the game to be played without saving files
#474
Added a mode called GAN to the icecreamyou code to be able to get levels without saving files and provided this funciton by altering constructors and getters in GamePanel, Level, and LodeRunner.
LodeRunnerVGLCUtil now lauches level1 into the game with and is playable.
Next going to do the same thing with a level from the gan as my next step.