#473 all 10 levels
Some blocks are not represented in the VGLC text files (some breakable blocks are not represented). Additionally, blocks that look solid but you fall through are simply empty spaces. Finally, VGLC does not contain boss rooms, so there will be no spawn area for them when we try and add that element to the game.

**NOTE** you have to put these files into %localappdata% under MegaManMaker/Levels to play