JUnit tests for Lode Runner Level 1 and Level 2
#501
LodeRunnerState class now has JUnit tests to test the actionSequence for both Level 1 and Level 2 of Lode Runner.
First the size of the actionSequence list is confirmed and that it is not null.
Then every action in the sequence is checked with an iterator.
Finally we check that we are at the end of the list after iterating through the whole thing.