Added connectivity fitness function
#508
Added a fitness function in LodeRunnerLevelTask that is the amount of ground that gets covered in the level during the search.
Updated default GAN model for lode runner to be from the first 100 levels with the 7 tile mapping.
#474
Updated MMNEAT to be able to take instances of LodeRunnerGANLevelTask.
 Added a Main method in LodeRunnerGANLevelTask
Started to write debugging code for LodeRunnerLevelTask in the if(CommonConstants.watch) if statement at the end of OneEval method