Added Parameters to control when fitness functions are used
#508
Added boolean parameters to control if simpleAStarPathDistance and connectivityOfLevel are used. Also updated variable names to be more descriptive.
Added control to only add the values for those varaibles if the boolean parameter if true.
Added the otherScores array that holds both of those values regardless of if they are used as fitness functions. @schrum2