Added and caluclated values for otherScores array
#513
Added 6 values that are not registered as fitness functions that are calculated in the evaluate method and passed into otherScores.
Wrote code to calculate the percent of ladders, the percent of ground, the percent of ropes the percent of connectivity, and the number of treasures and enemies
Also added a print out of those values for debugging in the if(CommonConstants.watch) debugging code. @schrum2