#514 #508 Actually set fitness functions true in batch file
Rather than assume the default values are true. Also, using %1 for command line params now since this seems to work.

This means that if @kdsteckel wants to use the file on his machine, he will have to repalce the %1 with $1 or hard code numbers, but this is how the batch file should remain.