Work on implementing MAPElites for LodeRunnerLevelTask
#513
Added code to define bin dimensions for levels that are generated.
Wrote a helper method to set the bins adn save a normal rendered image of the level and the rendered solution path. This is still not saving in the archieve directory correctly, but I think I am close.
Updated the main method of LodeRunnerGANLevelTask to use MAPElites.
@schrum2