MAPElites now puts images of levels and solution paths in achieve directory
#513
Updated main method of LodeRunnerGANLevelTask to have ea have MAPElites as a parameter, and set the experiement to a SteadyStateExperiment.
Changed the method, by using createBufferedImage instead of getBufferedImage, for the standard level render so that it just makes the image instead of displaying it in a window.
Updated the soution image save so that it actual saves the Solution instead of just a copy of the standard rendered level as it did before.
@schrum2