Fixed image file names
#513
Fixed the names of the files with the images so that they are located togther when sorted by filename.
Still need to figure out how to change the -score.txt and -elite.xml to have the genomeID and "_Direct" so that they appear next to the images they are associated with when sorting by filename.