Render to IceCreamYou in Interactive Breeding
#527
Wrote code in LodeRunnerLevelUtil to render an image of an IceCreamYou level to put into the level breeder, need to make the background white instead of black.
Added check box in the level breeder to switch to ice cream you visualization.
Added a parameter to control if the icecreamyou visualization is displayed in the level breeder.
Also added a few comments in the LodeRunnerLevelAnaylsisutil class #525