#528 Need better heuristic
This code demonstrates that if you reduce the number of treasures in Level 4 to 9, you can solve the game with a budget of 150000, but fail with a budget of 145000. If the number of treasures is increased to its proper number, then the required budget is infeasibly large.

I don't think getSuccessor is wrong ... I think we have to have a better heuristic