Set up experiment class to look at MAPElites levels
#541
Set up the experiment class for this, next I need to write the batch file to run it.
Added a string parameter witht he file name.
Fixed bug that put 4 spawn points in every level generated with MAPElites, just changed it so that it only performed the A* search once.
@schrum2