#528 More lenient A* model
@kdsteckel This new model allows for sideways movement through diggable blocks with certain restrictions and at high cost. Hopefully, this will make all levels beatable.

However, I now worry that some paths will be legal that shouldn't be. In fact, I added a constant LodeRunnerState.ALLOW_WEIRD_SIDE_WAYS_MOVE_THROUGH_DIGGABLE so that we can turn this feature on and off easily. I think more work is needed on it, but in the mean time, see if there are any levels that still can't be beaten using this more lenient movement model