#528 Sometimes, you can't add an edge
@kdsteckel The problem with level 41 seems to be that you have to collect the treasures in the upper right in a certain order. First you get the one to your left, then you use the rope to climb over to the right, then go down the stairs to the left.

However, if you drop of the rope early and get the treasure at the bottom of those stairs first, you're screwed ... it's impossible to get back up to the top of those stairs. Therefore, you legitimately should not have edges from the lower steps leading to the upper steps. That's why we do not add edges for a null result in those cases (thoush we still need to incorporate the code that tries solving things with then without weird moves).

However, using a proper TSP solution method should still be able to handle this. The problem is that the TSP solver hits a dead end where other treasures can't be reached.  I think the way to fix this level is to tweak the TSP solver instead.

In the meantime, check out the other levels and report on problems you find.