#515 All levels in the sequence are unique
Fixed the size of the genotype so that it would allow for individual levels in MMNEAT.
Changed getLevelSequence in LodeRunnerGANLevelTask to make it about to get the phenotype for each level individually.
Updated the naming of the fitness functions and other functions, but this still needs more work.
#528
Added a inBounds check in the up case of the getSuccessor method in LodeRunnerState.