#528 Troubleshooting Level 47
This does not work. Somehow, the graph is trying to remove a node that has already been removed.

Note: I believe I fixed a previous issue in the look-up of a missing Node by making sure that a copy of the original TSP was made and maintained through execution, but this simply changed when an error occurred.