#528 Finally solved Level 47
Ok ... maybe this time it is fixed. I rewrote the deepCopy method and hopefully made many other fixes. Level 47 can be solved ... but it's really slow!

I think that doing a topological sort on the graph could speed things up, since that really defines the order it needs to be searched in.