Work on integrating TSP into our code
#559
Added a drop down menu to decide if the user wants to use display A* path or the TSP path. Still needs work.
#525
Added the TSP path length as one of the things we are measuring in the level anaylsis.
#559
Added code to track the amount of back tracking the TSP solver needs to do to restrict to a budget.
#559
Added parameters for TSP budget, and to control the visualization in the level breeder.
Can you look at these when you get a chance? @schrum2