More work on increasing difficulty fitnesses functions
#519
Worked on getting the fitness functions for increasing difficulty in the level sequence to work.
Wrote method to find the values of the fitness function from other scores.
Wrote method to get the meanSquared errors for the fitnesses and the expected values.
Filled fitness array with the fitiness values.
Still needs more work
@schrum2