#519 Adjusted line definition based on min and max
Rather than treat the first and last fitness values as the endpoints of the line, we now treat the min and max values as the y values for the line endpoints. This change punishes level sequences along the "fit" objectives that do not have an increasing sequence ... before, this was being exclusively enforced by the "range" objectives, which was allowing weird solutions to have high fitness.