#565 More work on CPPN2GAN
I ended up just copying the code from MegaManGANUtil.longVectorToMegaManLevel for CPPN based generation. The only thing that changes is where the vector for one segment comes from. This is convenient, but need to be refactored/fixed eventually. Once we know it works, we can work on refactoring the code to remove the duplication.

@cappsb