Make behaviorVector private in Score
In setting up alternate ways to define a MAP Elites bin, I decided to restrict access to the behaviorVector in Score, which is how the MAP Elites bin is currently specified. There is also a new method to get a given bin's score, which bases the result either on the behaviorVector, or on the new more efficient approach I am developing. The new approach is not used anywhere yet. This change doesn't seem to have broken any code yet. #579