#579 Fixing the new efficient MAP Elites
There were several bugs. Old version was scanning all scores in the behavior vector, which still makes sense for innovation engines. However, new efficient approach should only check one bin index. This required changing code both to fix an exception, and for the sake of efficiency (which is the whole point of this)