#579 Erros resolved ... seems to work
I'm testing with one Zelda Dungeon task, and the number of filled bins at each step is identical to the old version using the full behavior vector. Need to run longer to test more, and also make similar changes to Mario, Lode Runner, and Mega Man.

However, before doing that, I may change the code yet again so that the complicated row-major order scaling doesn't have to happen in the game tasks. Currently, awareness of how to reduce the characterization to a bin in the 1D array is needed ... would make more sense for a MAP Elites user to only understand indices for the individual dimensions.