#579 Send multiple indices instead of the 1D index
It doesn't make sense for the user to have to use the row-major order formula each time a bin index is defined. In the task, just provide the indices for each dimension, and then have the mapping be responsible for knowing how to convert to row-major order.

This has been done with all of the mappings, but currently only Zelda actually uses the more efficient MAP Elites approach. Comparison tests have been done for the Zelda bin mapping schemes, and results are the same as with traditional behavior vectors.