#560 First attempt at LevelNovelty
Made a few edits to the body of DungeonNovelty to account for non-dungeon layouts. I added a parseLevel method that takes in a level and returns a list of all of its segments (except null segments). From there, I created a new class MegaManLevelNovelty that extends LevelNovelty and allows for more specific novelty testing (OneGAN, SevenGAN, original VGLC). Take a look @schrum2