#583 Fixed issue with bin directories
I found an old copy of VGG16 still on my machine, so I was able to test out the innovation engine code to see if it still runs. I actually had to fix an issue with file save paths by making sure that appropriate directories are created before evolved images get saved to them.

I can only do this with VGG16 since it happens to sill be on this machine, but I haven't really looked into fixing how to get the model weights in general.