#158 Just copy the DL4J version of Neural Style Transfer
I gave up on trying to reform my old code for NST to work. I don't know why the images were becoming white washed out.

So, instead I've shifted to just copying the DL4J code, and modifying it to work the way I need, which basically just depends on the API. I'm not done, and there are errors now. The Old version of NeuralStyleTransfer is still around, but I am making progress