#158 Found problem
The part of the code creating the combination image was loading the fixed content image from the file path instead of the specified image. Fixed now. So, DL4J NST works. It's really slow, but it works. Need to see if I can use CUDA with DL4J, but I'm going to close this issue.

Also need to fix Python NST as well ... new issue?