#583 Fixed Innovation Engine for DL4J
The Innovation using DL4J works again. The problem was the presence of multiple maven entries for the commons-codec. There was one version for RL Glue and another version that I added recently for DL4J. However, I removed both, so that DL4J could simply pull in whatever is needed as a dependency of something else rather than a top-level dependency.