#348 Jackson Serialization
Started work on Jackson based serialization method. I tested it with a very simply class, and that seemed to work, but it doesn't work with this TestObject that was originally part of wox. I'm not sure I actually need anything as complicated as this TestObject, but the fact that I'm losing functionality does bother me. At the very least, I need to figure out exactly what the source of the problem is. The stack trace is too long for me to figure out the original problem, but that makes me wonder if there is a circular reference problem, since wox brags about being able to handle these.