#611 Added another zentangleImages method with an extra parameter, BufferedImage pattern4
This new method now differentiates from the zentangleImages method that only has three BufferedImag parameters.  In the second case (one black and one non-black), it now matters which backgroundImage is black and which one is non-black.  In the case that the backgroundImage1 is black and backgroundImage2 is non-black, use the rgb from pattern2.  In the case backgroundImage1 is non-black and backgroundImage2 is black, use the rgb from pattern4.