#647 Added the new binning scheme to PictureTargetTask.
The new binning scheme has been added to PictureTargetTask as a part of an if else.  One if if we are using the neurons and links binning scheme, one if else if we are using the new binning scheme (PictureFourQuadrantBrightnessBinLabels), and an else that throws an IllegalStateException.  The array indicesMapEliteBin is set to hold all the indices of the bins of the image.