Made Pogamut libraries local
I'm getting an annoying error with jfrog #672 and I thought it was related to the Pogamut libraries. I thought that if I switched them all from being downloaded to coming from the local repo included in the github repo (in libs directory) that it would solve the problem. The local libraries seem to work, but now I'm getting another error from jfrog! It makes me wonder if I should just leave the Pogamut libraries the way they are, but it was quite a hassle to get them working from the local repo, and I probably won't need to constantly get up-to-date versions anyway. In fact, I have a legitimate concern that the Pogamut libraries could disappear, so maybe keeping them local is a good idea. In any case, it will be easy to undo this commit later if I need to.