#669 Moved/renamed/changed batch file
It makes more sense for this batch file to exist in the main MM-NEAT directory since it is a post-experiment analysis tool. I also generalized the name, and made the other binning scheme a command line parameter, even though that will generally be rough to type out ... but we'll copy-paste anyway.