#642 Fixes so AutoEncoder can initialize
There is a special case when starting up the autoencoder. The initial images need to be generated and evaluated, but without having an autoencoder to base their score on.

The approach I used is to give them all the same autoencoder score (loss of 1.0), which is problematic since that means there will only be one initial image. However, the CPPN Complexity Mapping has the same handicap.

Really need to find a fix that is fair for both of them.