#646 Changed -Infinity to X in logs
This saves space more so than time, but it is still a good change to make. I also updated some Python scripts to look for -Infinity instead of X. There were also lots of R scripts that I thought would be affected, but apparently not ... either the X is being treated the same way as the -Infinity, or those parts of the scripts weren't really doing what I thought they were in the first place. I only tried out one, but it worked. In all of them, I made a note in the comments in case problems arise later.

@MBatt1 Note these changes, but be aware that I made them in a new branch dev_schrum. For now, just ignore this branch, but I'm going to keep working on optimizations here. My only worry is that you can't safely merge them without making your already collected results for #686 incompatible. That's why you should avoid the branch for now.