#646 Removed lower/upper bounds form Bounded Genotype
There was no need for every copy of a bounded genotype to store a copy of the lower and upper bounds. On top of that, there was no reason to save that data to every xml file for each genotype (if the task is initialized the same way, then MMNEAT should sets the bounds appropriately).

However, removing the bounds surprisingly required many more changes than expected. The code Maven builds, but I definitely worry about uncaught bugs now.